cleanup PostgresMessageType
